content-data-control-center/requirements.txt
anthropic
streamlit
gspread
google-auth
google-auth-oauthlib
google-api-python-client
requests
beautifulsoup4
python-dotenv
